Co-locate backend test suites with feature slices
