+++ app/cli/tests/test_app.py
@@ -1,0 +1,81 @@
+"""CLI parser wiring tests."""
+
+from __future__ import annotations
+
+from app.cli.commands import api_keys, users
+from app.cli.commands import reset as reset_command
+from app.cli.commands import settings as settings_command
+from app.cli.main import build_cli_app
+
+
+def test_users_create_handler_resolution() -> None:
+    parser = build_cli_app()
+    args = parser.parse_args(
+        [
+            "users",
+            "create",
+            "--email",
+            "example@example.test",
+            "--password",
+            "secret-value",
+        ]
+    )
+    assert args.handler is users.create
+
+
+def test_users_activate_accepts_email_identifier() -> None:
+    parser = build_cli_app()
+    args = parser.parse_args(
+        [
+            "users",
+            "activate",
+            "--email",
+            "example@example.test",
+        ]
+    )
+    assert args.handler is users.activate
+    assert args.email == "example@example.test"
+    assert args.user_id is None
+
+
+def test_api_keys_issue_handler_resolution() -> None:
+    parser = build_cli_app()
+    args = parser.parse_args(
+        [
+            "api-keys",
+            "issue",
+            "--user-id",
+            "user-123",
+        ]
+    )
+    assert args.handler is api_keys.issue
+
+
+def test_start_command_collects_env_overrides() -> None:
+    parser = build_cli_app()
+    args = parser.parse_args(
+        [
+            "start",
+            "--env",
+            "ADE_LOGGING_LEVEL=DEBUG",
+            "--env",
+            "ADE_API_DOCS_ENABLED=true",
+        ]
+    )
+
+    assert args.env == ["ADE_LOGGING_LEVEL=DEBUG", "ADE_API_DOCS_ENABLED=true"]
+
+
+def test_settings_command_handler_resolution() -> None:
+    parser = build_cli_app()
+    args = parser.parse_args(["settings"])
+
+    assert args.handler is settings_command.dump
+
+
+def test_reset_command_handler_resolution() -> None:
+    parser = build_cli_app()
+    args = parser.parse_args(["reset", "--yes"])
+
+    assert args.handler is reset_command.reset
+    assert args.yes is True
